requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
requests==2.32.3
beautifulsoup4==4.12.3
python-jose[cryptography]
passlib[bcrypt]
APScheduler==3.10.4
pytest==8.3.3
black==24.3.0
flake8==7.0.0
email-validator==2.1.0
bcrypt==4.1.1
redis==5.2.1